Finalisation du service Facturation et debut du service Commande

+++ Facturation/src/main/java/com/openlab/facturation/dto/FactureDTO.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class FactureDTO {
     private Long id;
     private Long commandeId;
-    private Long tableNumero;
+    private int tableNumero;
     private double montantTotal;
     private LocalDateTime dateFacture;
 }
